backend/requirements.txt
# ── Google Gemini 3 (official new SDK) ────────────────────────
google-genai>=1.33.0

# ── LangGraph (agent state machines) ─────────────────────────
langgraph>=0.2.45
langchain>=0.3.7
langchain-google-genai>=2.0.4
langchain-community>=0.3.7

# ── RAG: ChromaDB vector store + sentence embeddings ─────────
chromadb>=0.5.18
sentence-transformers>=3.2.1

# ── API server ────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
websockets>=13.1
python-multipart>=0.0.12

# ── Utilities ─────────────────────────────────────────────────
python-dotenv>=1.0.1
pydantic>=2.9.2
numpy>=1.26.4
